dev_01_UIApproach  - Home approach. Delete Splash approach.
