+++ lib/main.dart
@@ -1,15 +1,13 @@
 import 'package:flutter/material.dart';
-import 'screens/splash.dart';
+import 'package:coloweather/screens/homepage.dart';
+import 'package:intl/date_symbol_data_local.dart';
 
-void main() {
-  runApp(MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Splash(),
-    );
-  }
+main(List<String> args) {
+  initializeDateFormatting('pt_BR', null).then((_) {
+    runApp(MaterialApp(
+      home: HomePage(),
+      theme: ThemeData.light(),
+      debugShowCheckedModeBanner: false,
+    ));
+  });
 }
